separate create logic to create model
